fix errors in controllers

+++ src/controllers/comments.controller.ts
@@ -5,20 +5,36 @@ export async function createComment(
   req: Request,
   res: Response,
 ): Promise<void> {
-  const comment = await prisma.comment.create({
-    data: {
-      content: req.body.content,
-      postId: parseInt(req.params.id),
-      authorId: req.body.user.id,
-    },
-  })
-  res.status(200).json({ comment: comment })
+  try {
+    const comment = await prisma.comment.create({
+      data: {
+        content: req.body.content,
+        postId: parseInt(req.params.id),
+        authorId: req.body.user.id,
+      },
+    })
+    res.status(200).json({ comment: comment })
+  } catch (error) {
+    return res
+      .status(404)
+      .json({ error: `Couldn't find comment with id = '${req.params.id}'` })
+      .end()
+  }
 }
 
 export async function findComments(req: Request, res: Response): Promise<void> {
   const allComments = await prisma.comment.findMany({
     where: { postId: parseInt(req.params.id) },
   })
+
+  if (!allComments.length)
+    return res
+      .status(404)
+      .json({
+        error: `Couldn't find comments with post id = '${req.params.id}'`,
+      })
+      .end()
+
   res.status(200).json({ comments: allComments })
 }
 
@@ -28,7 +44,10 @@ export async function findComment(req: Request, res: Response): Promise<void> {
   })
 
   if (!comment)
-    throw new Error(`Couldn't find post with id = '${req.params.id}'`)
+    return res
+      .status(404)
+      .json({ error: `Couldn't find comment with id = '${req.params.id}'` })
+      .end()
 
   res.status(200).json({ comment: comment })
 }
@@ -37,7 +56,14 @@ export async function updateComment(
   req: Request,
   res: Response,
 ): Promise<void> {
-  await errorCommentUserIdDiff(req, res)
+  const permissionError = await errorCommentUserIdDiff(req)
+  if (permissionError)
+    return res
+      .status(403)
+      .json({
+        error: `Current user cannot edit the information of another user, only that of him`,
+      })
+      .end()
   const commentBody = { ...req.body }
   delete commentBody.user
 
@@ -48,7 +74,10 @@ export async function updateComment(
     })
     res.status(200).json({ comment: comment })
   } catch (error) {
-    throw new Error(`Couldn't find comment with id = '${req.params.id}'`)
+    return res
+      .status(404)
+      .json({ error: `Couldn't find comment with id = '${req.params.id}'` })
+      .end()
   }
 }
 
@@ -56,7 +85,14 @@ export async function deleteComment(
   req: Request,
   res: Response,
 ): Promise<void> {
-  await errorCommentUserIdDiff(req, res)
+  const permissionError = await errorCommentUserIdDiff(req)
+  if (permissionError)
+    return res
+      .status(403)
+      .json({
+        error: `Current user cannot edit the information of another user, only that of him`,
+      })
+      .end()
   try {
     const comment = await prisma.comment.delete({
       where: { id: parseInt(req.params.id) },
@@ -64,7 +100,10 @@ export async function deleteComment(
 
     res.status(200).json({ comment: comment })
   } catch (error) {
-    throw new Error(`Couldn't find comment with id = '${req.params.id}'`)
+    return res
+      .status(404)
+      .json({ error: `Couldn't find comment with id = '${req.params.id}'` })
+      .end()
   }
 }
 
@@ -72,7 +111,14 @@ export async function publishComment(
   req: Request,
   res: Response,
 ): Promise<void> {
-  await errorCommentUserIdDiff(req, res)
+  const permissionError = await errorCommentUserIdDiff(req)
+  if (permissionError)
+    return res
+      .status(403)
+      .json({
+        error: `Current user cannot edit the information of another user, only that of him`,
+      })
+      .end()
   try {
     const comment = await prisma.comment.update({
       where: { id: parseInt(req.params.id) },
@@ -80,7 +126,10 @@ export async function publishComment(
     })
     res.status(200).json({ comment: comment })
   } catch (error) {
-    throw new Error(`Couldn't find comment with id = '${req.params.id}'`)
+    return res
+      .status(404)
+      .json({ error: `Couldn't find comment with id = '${req.params.id}'` })
+      .end()
   }
 }
 
@@ -126,6 +175,9 @@ export async function likedComment(req: Request, res: Response): Promise<void> {
     }
     res.status(200).json({ sucess: true })
   } catch (error) {
-    throw new Error(`Couldn't find comment with id = '${req.params.id}'`)
+    return res
+      .status(404)
+      .json({ error: `Couldn't find comment with id = '${req.params.id}'` })
+      .end()
   }
 }
