+++ src/controllers/posts.controller.ts
@@ -2,18 +2,29 @@ import { Request, Response } from 'express'
 import { errorPostUserIdDiff, prisma } from './app.controller'
 
 export async function createPost(req: Request, res: Response): Promise<void> {
-  const post = await prisma.post.create({
-    data: {
-      title: req.body.title,
-      content: req.body.content,
-      authorId: req.body.user.id,
-    },
-  })
-  res.status(200).json({ post: post })
+  try {
+    const post = await prisma.post.create({
+      data: {
+        title: req.body.title,
+        content: req.body.content,
+        authorId: req.body.user.id,
+      },
+    })
+    res.status(200).json({ post: post })
+  } catch (error) {
+    return res
+      .status(404)
+      .json({ error: `Couldn't find pos with id = '${req.params.id}'` })
+      .end()
+  }
 }
 
 export async function findPosts(req: Request, res: Response): Promise<void> {
   const allPosts = await prisma.post.findMany()
+
+  if (!allPosts)
+    return res.status(404).json({ error: `Couldn't find any posts jet ` }).end()
+
   res.status(200).json({ posts: allPosts })
 }
 
@@ -22,13 +33,25 @@ export async function findPost(req: Request, res: Response): Promise<void> {
     where: { id: parseInt(req.params.id) },
   })
 
-  if (!post) throw new Error(`Couldn't find post with id = '${req.params.id}'`)
+  if (!post)
+    return res
+      .status(404)
+      .json({ error: `Couldn't find pos with id = '${req.params.id}'` })
+      .end()
 
   res.status(200).json({ post: post })
 }
 
 export async function updatePost(req: Request, res: Response): Promise<void> {
-  await errorPostUserIdDiff(req, res)
+  const permissionError = await errorPostUserIdDiff(req)
+  if (permissionError)
+    return res
+      .status(403)
+      .json({
+        error: `Current user cannot edit the information of another user, only that of him`,
+      })
+      .end()
+
   const postBody = { ...req.body }
   delete postBody.user
 
@@ -39,12 +62,22 @@ export async function updatePost(req: Request, res: Response): Promise<void> {
     })
     res.status(200).json({ post: post })
   } catch (error) {
-    throw new Error(`Couldn't find post with id = '${req.params.id}'`)
+    return res
+      .status(404)
+      .json({ error: `Couldn't find pos with id = '${req.params.id}'` })
+      .end()
   }
 }
 
 export async function deletePost(req: Request, res: Response): Promise<void> {
-  await errorPostUserIdDiff(req, res)
+  const permissionError = await errorPostUserIdDiff(req)
+  if (permissionError)
+    return res
+      .status(403)
+      .json({
+        error: `Current user cannot edit the information of another user, only that of him`,
+      })
+      .end()
   try {
     const post = await prisma.post.delete({
       where: { id: parseInt(req.params.id) },
@@ -52,12 +85,22 @@ export async function deletePost(req: Request, res: Response): Promise<void> {
 
     res.status(200).json({ post: post })
   } catch (error) {
-    throw new Error(`Couldn't find post with id = '${req.params.id}'`)
+    return res
+      .status(404)
+      .json({ error: `Couldn't find pos with id = '${req.params.id}'` })
+      .end()
   }
 }
 
 export async function publishPost(req: Request, res: Response): Promise<void> {
-  await errorPostUserIdDiff(req, res)
+  const permissionError = await errorPostUserIdDiff(req)
+  if (permissionError)
+    return res
+      .status(403)
+      .json({
+        error: `Current user cannot edit the information of another user, only that of him`,
+      })
+      .end()
   try {
     const post = await prisma.post.update({
       where: { id: parseInt(req.params.id) },
@@ -65,7 +108,10 @@ export async function publishPost(req: Request, res: Response): Promise<void> {
     })
     res.status(200).json({ post: post })
   } catch (error) {
-    throw new Error(`Couldn't find post with id = '${req.params.id}'`)
+    return res
+      .status(404)
+      .json({ error: `Couldn't find pos with id = '${req.params.id}'` })
+      .end()
   }
 }
 
@@ -111,6 +157,9 @@ export async function likedPost(req: Request, res: Response): Promise<void> {
     }
     res.status(200).json({ sucess: true })
   } catch (error) {
-    throw new Error(`Couldn't find post with id = '${req.params.id}'`)
+    return res
+      .status(404)
+      .json({ error: `Couldn't find pos with id = '${req.params.id}'` })
+      .end()
   }
 }
